tipo de login acceso de datos

+++ views/clientes/index.php
@@ -1,3 +1,12 @@
+<?php
+error_reporting(error_reporting() & ~E_NOTICE);
+if (session_status() == PHP_SESSION_NONE) {
+    session_start();
+}
+if (isset($_SESSION['profile'])) {
+    $profile = $_SESSION['profile'];
+}
+?>
 <!-- Content Wrapper. Contains page content -->
 <div class="content-wrapper">
     <!-- Content Header (Page header) -->
@@ -52,7 +61,9 @@
                                     <div class="btn-group">
                                         <a class="btn btn-warning" href="/clientes/actualizar?id=<?php echo $client->id; ?>"><i class="fa fa-edit"></i></a>
 
-                                        <a class="btn btn-danger avisoAlertaxx" href="/clientes/eliminar?id=<?php echo $client->id; ?>&tipo=client"><i class="fa fa-times"></i></a>
+                                        <?php if ($profile == "Administrador") : ?>
+                                            <a class="btn btn-danger avisoAlertaxx" href="/clientes/eliminar?id=<?php echo $client->id; ?>&tipo=client"><i class="fa fa-times"></i></a>
+                                        <?php endif; ?>
                                     </div>
                                 </td>
                             </tr>
